SbAudio: Remove XACT
This is very old and highly specific technology that's hard to use today,
hard to find documentation for, etc. At least it seems this way.

FAudio agrees: "This file has no documentation since you are expected to already know how XACT works if you are still using these APIs!

For instance wavebanks and soundbanks are authored in tools I have no access to,
sit on disk, and are parsed by this XACT code. The code is thus not very useful
if we have no way to author these files.

There are however some good ideas in here we can learn from and provide
in a modern solution.