SbAudio: Remove libsoundio dependency
Change of plans (this had not yet been integrated)

libsoundio is a high quality library that due to its quality I wanted to try integrating. It's not a bad idea to do so, but it's a generic low-level wrapper around platform audio APIs. We don't need generic. SbAudio is just intended to support one API per platform for now, and it only needs the backend functionality needed to implement what it provides.

libsoundio also requires binary attribution and FAudio doesn't. FAudio already provides a WASAPI backend so all that's left is a PulseAudio or Pipewire backend. This will be custom written and SDL thrown out. No binary attribution requirement added.

The real change required for SbAudio is getting rid of the SDL dependency.