Port: [win32] List all audio devices for Windows (#338)
Previously, FAudio with `PLATFORM_WIN32` would only list the default audio device. This reworks the FAudio `PLATFORM_WIN32` code to enable access to every audio device on the system. In order to preserve the existing behavior (and XAudio2 compabitility), the audio devices are re-ordered so the first device is always the default one.

In addition:

* Properly populates the `DisplayName` field in `FAudioDeviceDetails`. Previously, it was set to the device GUID.
* [windows] Return error code for DefaultDeviceIndex
* Rework `FAudio_DefaultDeviceIndex` to return the Windows API error code and take the `defaultDeviceIndex` as parameter.

Test: Local build with visualboyadvance-m