Port: Validate PlayBegin+PlayLength against AudioBytes.
I have a game that submits garbage data and crashes. My Ad-hoc testing
shows that Xaudio2.8 and 2.9 validate the length, but XAudio2.7 does
not. I tried to write tests, but I ran into multiple issues with
tests/xaudio2.c and the file doesn't seem to be touched since 5 years.

---

List of issues with the tests from memory:

* IXAudio27 don't exist any more in modern Wine headers. Wine 8.0 has them
  though.

* Xaudio2.8 doesn't detect any devices, GetDeviceCount returns garbage.
  AFAIU it isn't supposed to be used for 2.8 and newer.

* If I hack around that, IXAudio2MasteringVoice::SetEffectChain crashes.

* If I disable that, IXAudio2SourceVoice::GetState crashes on xaudio2.8.

At that point I gave up and decided to submit this without tests.